perf!: client / server subpaths
