--- packages/dom/README.md
@@ -1,12 +1,0 @@
-# @unhead/dom
-
-## Install
-
-```bash
-# yarn add @unhead/dom
-npm install @unhead/dom
-```
-
-## Documentation
-
-See the [@unhead/dom](https://unhead.unjs.io/guide/getting-started/how-it-works#unheaddom) for how this works.
