--- packages/dom/export-size-report.json
@@ -1,36 +1,0 @@
-{
-  "meta": {
-    "name": "@unhead/dom",
-    "dependencies": [
-      "@unhead/schema",
-      "@unhead/shared"
-    ],
-    "versions": {
-      "export-size": "0.7.0",
-      "esbuild": "^0.19.5"
-    }
-  },
-  "exports": [
-    {
-      "name": "DomPlugin",
-      "path": "dist/index.mjs",
-      "minified": 3126,
-      "minzipped": 1329,
-      "bundled": 5210
-    },
-    {
-      "name": "debouncedRenderDOMHead",
-      "path": "dist/index.mjs",
-      "minified": 2880,
-      "minzipped": 1222,
-      "bundled": 4763
-    },
-    {
-      "name": "renderDOMHead",
-      "path": "dist/index.mjs",
-      "minified": 2692,
-      "minzipped": 1158,
-      "bundled": 4429
-    }
-  ]
-}
