+++ packages/dom/package.json
@@ -29,12 +29,10 @@
     "dist"
   ],
   "scripts": {
-    "build": "unbuild .",
-    "stub": "unbuild . --stub",
-    "export:sizes": "npx export-size . -r"
+    "build": "unbuild ",
+    "stub": "unbuild  --stub"
   },
-  "dependencies": {
-    "@unhead/schema": "workspace:*",
-    "@unhead/shared": "workspace:*"
+  "peerDependencies": {
+    "unhead": "workspace:*"
   }
 }
