+++ packages/dom/src/index.ts
@@ -1,3 +1,1 @@
-export * from './debounced'
-export * from './domPlugin'
-export * from './renderDOMHead'
+export * from 'unhead/client'
