--- packages/ssr/README.md
@@ -1,12 +1,0 @@
-# @unhead/ssr
-
-## Install
-
-```bash
-# yarn add @unhead/dom
-npm install @unhead/ssr
-```
-
-## Documentation
-
-See the [Unhead](https://unhead.unjs.io/) for more details.
