--- packages/ssr/export-size-report.json
@@ -1,50 +1,0 @@
-{
-  "meta": {
-    "name": "@unhead/ssr",
-    "dependencies": [
-      "@unhead/schema",
-      "@unhead/shared"
-    ],
-    "versions": {
-      "export-size": "0.7.0",
-      "esbuild": "^0.19.5"
-    }
-  },
-  "exports": [
-    {
-      "name": "renderSSRHead",
-      "path": "dist/index.mjs",
-      "minified": 1542,
-      "minzipped": 656,
-      "bundled": 2665
-    },
-    {
-      "name": "ssrRenderTags",
-      "path": "dist/index.mjs",
-      "minified": 1182,
-      "minzipped": 539,
-      "bundled": 2059
-    },
-    {
-      "name": "tagToString",
-      "path": "dist/index.mjs",
-      "minified": 752,
-      "minzipped": 371,
-      "bundled": 1340
-    },
-    {
-      "name": "propsToString",
-      "path": "dist/index.mjs",
-      "minified": 248,
-      "minzipped": 171,
-      "bundled": 502
-    },
-    {
-      "name": "escapeHtml",
-      "path": "dist/index.mjs",
-      "minified": 249,
-      "minzipped": 136,
-      "bundled": 504
-    }
-  ]
-}
