+++ packages/ssr/src/index.ts
@@ -1,3 +1,1 @@
-export * from './renderSSRHead'
-export * from './util'
-export type { SSRHeadPayload } from '@unhead/schema'
+export * from 'unhead/server'
