+++ packages/unhead/build.config.ts
@@ -9,5 +9,7 @@ export default defineBuildConfig({
   entries: [
     { input: 'src/index', name: 'index' },
     { input: 'src/optionalPlugins/index', name: 'optionalPlugins' },
+    { input: 'src/server/index', name: 'server' },
+    { input: 'src/client/index', name: 'client' },
   ],
 })
