+++ packages/unhead/src/server/renderSSRHead.ts
