+++ packages/vue/build.config.ts
@@ -10,5 +10,7 @@ export default defineBuildConfig({
   entries: [
     { input: 'src/index', name: 'index' },
     { input: 'src/components/index', name: 'components' },
+    { input: 'src/server/index', name: 'server' },
+    { input: 'src/client/index', name: 'client' },
   ],
 })
