+++ packages/vue/src/createHead.ts
@@ -1,13 +1,9 @@
-import type { CreateHeadOptions, MergeHead } from '@unhead/schema'
 import type { Plugin } from 'vue'
-import type { MaybeComputedRef, ReactiveHead, VueHeadClient } from './types'
-import { createHead as _createHead, createServerHead as _createServerHead } from 'unhead'
-import { nextTick } from 'vue'
-import VueReactivityPlugin from './plugins/VueReactivityPlugin'
+import type { VueHeadClient } from './types'
 
 export const headSymbol = 'usehead'
 
-function vueInstall(head: VueHeadClient<any>) {
+export function vueInstall(head: VueHeadClient<any>) {
   const plugin = <Plugin> {
     install(app) {
       app.config.globalProperties.$unhead = head
@@ -17,20 +13,4 @@ function vueInstall(head: VueHeadClient<any>) {
     },
   }
   return plugin.install
-}
-
-export function createServerHead<T extends MergeHead>(options: Omit<CreateHeadOptions, 'domDelayFn' | 'document'> = {}): VueHeadClient<T> {
-  const head = _createServerHead<MaybeComputedRef<ReactiveHead<T>>>(options) as VueHeadClient<T>
-  head.use(VueReactivityPlugin)
-  head.install = vueInstall(head)
-  return head
-}
-
-// TODO rename to createDOMHead
-export function createHead<T extends MergeHead>(options: CreateHeadOptions = {}): VueHeadClient<T> {
-  options.domDelayFn = options.domDelayFn || (fn => nextTick(() => setTimeout(() => fn(), 0)))
-  const head = _createHead<MaybeComputedRef<ReactiveHead<T>>>(options) as VueHeadClient<T>
-  head.use(VueReactivityPlugin)
-  head.install = vueInstall(head)
-  return head
 }
