+++ packages/vue/test/e2e/basic.test.ts
@@ -3,7 +3,9 @@
 import type { ReactiveHead } from '@unhead/vue'
 import { renderDOMHead } from '@unhead/dom'
 import { renderSSRHead } from '@unhead/ssr'
-import { createHead, createServerHead, unheadCtx, useHead, useServerHead } from '@unhead/vue'
+import { unheadCtx, useHead, useServerHead } from '@unhead/vue'
+import { createHead } from '@unhead/vue/client'
+import { createHead as createServerHead } from '@unhead/vue/server'
 import { describe, it } from 'vitest'
 import { useDom } from '../../../../test/fixtures'
 import { csrVueAppWithUnhead, ssrVueAppWithUnhead } from '../util'
