+++ test/bench/ssr-harlanzw-com-e2e.bench.ts
@@ -2,7 +2,8 @@ import type { Head } from '@unhead/schema'
 import { InferSeoMetaPlugin } from '@unhead/addons'
 import { definePerson, defineWebPage, defineWebSite, UnheadSchemaOrg, useSchemaOrg } from '@unhead/schema-org'
 import { renderSSRHead } from '@unhead/ssr'
-import { createServerHead, unheadCtx, useHead, useSeoMeta, useServerHead } from 'unhead'
+import { createHead as createServerHead } from '@unhead/vue/server'
+import { unheadCtx, useHead, useSeoMeta, useServerHead } from 'unhead'
 import { bench, describe } from 'vitest'
 
 describe('ssr e2e bench', () => {
